각 포멧별로 구분함. editfootage-submit 추가함.
