+++ http_houdini.go
@@ -1,0 +1,125 @@
+package main
+
+import (
+	"fmt"
+	"io/ioutil"
+	"log"
+	"net/http"
+	"os"
+	"path/filepath"
+
+	"golang.org/x/sys/unix"
+)
+
+func handleAddHoudini(w http.ResponseWriter, r *http.Request) {
+	token, err := GetTokenFromHeader(w, r)
+	if err != nil {
+		http.Redirect(w, r, "/signin", http.StatusSeeOther)
+		return
+	}
+	w.Header().Set("Content-Type", "text/html")
+	err = TEMPLATES.ExecuteTemplate(w, "addhoudini", token)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+}
+
+func handleAddHoudiniProcess(w http.ResponseWriter, r *http.Request) {
+	token, err := GetTokenFromHeader(w, r)
+	if err != nil {
+		http.Redirect(w, r, "/signin", http.StatusSeeOther)
+		return
+	}
+	w.Header().Set("Content-Type", "text/html")
+	err = TEMPLATES.ExecuteTemplate(w, "addhoudini-process", token)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+}
+
+// handleAddHoudiniProcess 함수는 Houdini 파일을 처리하는 페이지 이다.
+func handleUploadHoudini(w http.ResponseWriter, r *http.Request) {
+	_, err := GetTokenFromHeader(w, r)
+	if err != nil {
+		http.Redirect(w, r, "/signin", http.StatusSeeOther)
+		return
+	}
+	file, header, err := r.FormFile("file")
+	if err != nil {
+		log.Println(err)
+	}
+	defer file.Close()
+	unix.Umask(0)
+	mimeType := header.Header.Get("Content-Type")
+	switch mimeType {
+	case "image/jpeg", "image/png":
+		data, err := ioutil.ReadAll(file)
+		if err != nil {
+			fmt.Fprintf(w, "%v", err)
+			return
+		}
+		path := os.TempDir() + "/dotori/thumbnail"
+		err = os.MkdirAll(path, 0770)
+		if err != nil {
+			return
+		}
+		err = ioutil.WriteFile(path+"/"+header.Filename, data, 0440)
+		if err != nil {
+			fmt.Fprintf(w, "%v", err)
+			return
+		}
+	case "video/quicktime", "video/mp4", "video/ogg", "application/ogg":
+		data, err := ioutil.ReadAll(file)
+		if err != nil {
+			fmt.Fprintf(w, "%v", err)
+			return
+		}
+		path := os.TempDir() + "/dotori/preview"
+		err = os.MkdirAll(path, 0770)
+		if err != nil {
+			return
+		}
+		err = ioutil.WriteFile(path+"/"+header.Filename, data, 0440)
+		if err != nil {
+			fmt.Fprintf(w, "%v", err)
+			return
+		}
+	case "application/octet-stream":
+		ext := filepath.Ext(header.Filename)
+		if ext == ".hip" || ext == ".hda" {
+			data, err := ioutil.ReadAll(file)
+			if err != nil {
+				fmt.Fprintf(w, "%v", err)
+				return
+			}
+			path := os.TempDir() + "/dotori"
+			err = os.MkdirAll(path, 0770)
+			if err != nil {
+				return
+			}
+			err = ioutil.WriteFile(path+"/"+header.Filename, data, 0440)
+			if err != nil {
+				fmt.Fprintf(w, "%v", err)
+				return
+			}
+		}
+	default:
+		data, err := ioutil.ReadAll(file)
+		if err != nil {
+			fmt.Fprintf(w, "%v", err)
+			return
+		}
+		path := os.TempDir() + "/dotori"
+		err = os.MkdirAll(path, 0770)
+		if err != nil {
+			return
+		}
+		err = ioutil.WriteFile(path+"/"+header.Filename, data, 0440)
+		if err != nil {
+			fmt.Fprintf(w, "%v", err)
+			return
+		}
+	}
+}
